requirements.txt
flask
pandas
google-genai
python-dotenv
reportlab
